use earliest balance instead of total balance for UG
