+++ mobile-money-receipts/main.js
@@ -1,10 +1,12 @@
-
+const getTranslation = require('./translations');
 var client = JSON.parse(contact.vars.client);
-var LoanBalance = 0;
+var TotalLoanBalance = 0;
+var earliestBalance = 0;
 var arrayLength = client.BalanceHistory.length;
 for (var i = 0; i < arrayLength; i++) {
     if (client.BalanceHistory[i].Balance > 0) {
-        LoanBalance = LoanBalance + client.BalanceHistory[i].Balance;
+        earliestBalance =client.BalanceHistory[i].Balance 
+        TotalLoanBalance = TotalLoanBalance + client.BalanceHistory[i].Balance;
     }
 }
 
@@ -14,7 +16,7 @@ var receipt = getTranslation('payment_receipt_ug', {
     balance: earliestBalance
 }, 'en')
 
-sendMessage({ 
+sendMessage({
     "to_number": contact.phone_number,
     "content": receipt
- });
+});
