+++ mobile-money-receipts/test-client-data.js
@@ -1,0 +1,69 @@
+const client = {
+    "GlobalClientId": "f456ff55-2d01-ea21-a42d-505bc2b76563",
+    "AccountNumber": "123456789",
+    "ClientName": "TCHALLA, PRINCE",
+    "FirstName": "PRINCE",
+    "LastName": "TCHALLA",
+    "CreatedDate": "2019-11-12T12:18:00.557",
+    "EnrollmentDate": "2019-11-12T12:18:00.557",
+    "BannedDate": null,
+    "DeceasedDate": null,
+    "EarliestCreatedDate": "2019-11-12T12:18:00.557",
+    "EarliestEnrollmentDate": "2019-11-12T12:18:00.557",
+    "LatestBannedDate": null,
+    "ClientId": 7201,
+    "GroupId": 678,
+    "GroupName": "BIKUWERAKI",
+    "SiteId": 22,
+    "SiteName": "Mafubira Buwekula",
+    "DistrictId": 2800,
+    "DistrictName": "Jinja",
+    "RegionId": 1800,
+    "RegionName": "South East",
+    "CountryId": 800,
+    "CountryName": "Uganda",
+    "AccountHistory": [
+        {
+            "AccountGuid": "f456ff55-2d01-ea21-a42d-505bc2b76563",
+            "ClientId": 7201,
+            "AccountNumber": "123456789",
+            "DistrictId": 2800,
+            "DistrictName": "Jinja",
+            "RegionId": 1800,
+            "RegionName": "South East",
+            "CountryId": 800,
+            "CountryName": "Uganda"
+        }
+    ],
+    "BalanceHistory": [
+        {
+            "AccountGuid": "f456ff55-2d01-ea21-a42d-505bc2b76563",
+            "GroupId": 678,
+            "GroupName": "BIKUWERAKI",
+            "SiteId": 22,
+            "SiteName": "Mafubira Buwekula",
+            "SeasonId": 270,
+            "SeasonName": "2020, Short Rain",
+            "SeasonStart": "2020-08-01T00:00:00",
+            "TotalCredit": 128000.000,
+            "TotalRepayment_IncludingOverpayments": 0.0000,
+            "Balance": 128000.0000,
+            "CurrencyCode": "UGX"
+        },
+        {
+            "AccountGuid": "f456ff55-2d01-ea21-a42d-505bc2b76563",
+            "GroupId": 678,
+            "GroupName": "BIKUWERAKI",
+            "SiteId": 22,
+            "SiteName": "Mafubira Buwekula",
+            "SeasonId": 260,
+            "SeasonName": "2020, Long Rain",
+            "SeasonStart": "2020-03-01T00:00:00",
+            "TotalCredit": 164000.000,
+            "TotalRepayment_IncludingOverpayments": 147000.0000,
+            "Balance": 17000.0000,
+            "CurrencyCode": "UGX"
+        }
+    ]
+};
+exports.client = client;
